+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 - Fix pagination parameters for `GetNextPageOfChildren` function for User service
+- Add `Surcharges` property to `Rate` class
 
 ## v6.7.0 (2024-07-24)